this changeset is for : 1. Making the hygieia jenkins plugin work with Sonar 5.2+ versions. 2. Capture new sonar metrics available from Sonar 5.0 onwards: new_violations, new_major_violations, new_critical_violations,new_blocker_violations, new_coverage, new_line_coverage, new_lines_to_cover    These new metrics capture metrics of only incremental code thus helping to fix "the leaky water pipe problem"
From Sonar 5.2+ onwards, SonarQube has hived off analysis of reports as an asynchronous process. This means that when Sonar analyss report is uploaded to Sonar during the sonar:sonar goal, the analysis does not get immediately completed, but Sonar sets an API endpoint called Compute Engine (CE) that gets updated when the corresponding analysis is available.

Hence, with Sonar 5.2+, the Jenkins job will need to monitor the CE API to determine if the processing is complete. After that is complete, can the Hygieia Publisher fetch the metrics from SonarQube for the project.

Also, the CE API can be queried only upto a maximum of 'X' attempts. This value is by default configured to 30 in SonarQube. For max attempts=30, Sonar recommends to make calls to CE API after every 10 seconds.

The changes involve:
1. Added additional configuration :
    i) Max attempts to CE API
    ii) Interval period between successive attempts to CE API
 This required modifying HygieiaPublisher and adding 2 html help files
 2. Modified SonarBuilder.java to:
    i) Determine CE API URL by inspecting the log from sonar analysis output
    ii) If CE API URL is found, it means the Sonar version is 5.2+. For this, keep polling CE API upto 'X' times with interval of 'Y' seconds.
    iii) If CE API provides a SUCCESS response within maximum number of attempts, continue processing the Sonar resources API
    iv) If CE API does not provide a SUCCESS response within maximum number of attempts or provides a failure response, then do not invoke the Sonar Resources API as the report wasn't available.
    v) If CE API URL is not found in the Sonar logs, it means the version is <5.2 and so the processing can continue as before.